% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iTunesR.R
\name{getAttributes}
\alias{getAttributes}
\title{Get Basic App Attributes from iTunes App Store}
\usage{
getAttributes(app_id, country)
}
\arguments{
\item{app_id}{A number Your iOS App ID for which reviews need to be downloaded}

\item{country}{A string country code like 'in' 'gb' for which reviews need to be downloaded}
}
\value{
A Data frame of App Attributes containing Title, Developer, Package Name, URL, Category
}
\description{
Get Basic App Attributes from iTunes App Store
}
