% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iTunesR.R
\name{getLogo}
\alias{getLogo}
\title{Download App Logo from iTunes App Store}
\usage{
getLogo(app_id, country)
}
\arguments{
\item{app_id}{A number Your iOS App ID for which reviews need to be downloaded}

\item{country}{A string country code like 'in' 'gb' for which reviews need to be downloaded}
}
\value{
downloads and saves App Logo in the current working directory
}
\description{
Download App Logo from iTunes App Store
}
\examples{
getLogo(742044692,'in')
}

