% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iTunesR.R
\name{getReviews}
\alias{getReviews}
\title{Extract Reviews from iTunes App Store}
\usage{
getReviews(app_id, country, page_num)
}
\arguments{
\item{app_id}{A number Your iOS App ID for which reviews need to be downloaded}

\item{country}{A string country code like 'in' 'gb' for which reviews need to be downloaded}

\item{page_num}{A number page number from 1 to 10 with 1 being the most-recent}
}
\value{
A Data frame of Reviews and meta information
}
\description{
Extract Reviews from iTunes App Store
}
\examples{
getReviews(742044692,'in',1)
}

